+++ app/components/folio/embed/box_component.rb
@@ -5,22 +5,22 @@ class Folio::Embed::BoxComponent < ApplicationComponent
                  data: nil,
                  centered: true,
                  background_color: nil,
-                 light_background_color: nil,
-                 dark_background_color: nil,
+                 light_mode_background_color: nil,
+                 dark_mode_background_color: nil,
                  class_name: nil)
     @folio_embed_data = folio_embed_data.is_a?(Hash) ? folio_embed_data : {}
     @data = data
     @centered = centered
     @background_color = background_color
-    @light_background_color = light_background_color
-    @dark_background_color = dark_background_color
+    @light_mode_background_color = light_mode_background_color
+    @dark_mode_background_color = dark_mode_background_color
     @class_name = class_name
   end
 
   private
     def before_render
       if dual_theme_background?
-        color = @light_background_color
+        color = @light_mode_background_color
         @style = "background-color: #{color};"
         @low_luminance = get_luminance(color) < 0.5
       elsif @background_color.present?
@@ -30,7 +30,7 @@ class Folio::Embed::BoxComponent < ApplicationComponent
     end
 
     def dual_theme_background?
-      @light_background_color.present? && @dark_background_color.present?
+      @light_mode_background_color.present? && @dark_mode_background_color.present?
     end
 
     def wrap_data
@@ -42,8 +42,8 @@ class Folio::Embed::BoxComponent < ApplicationComponent
                                 intersected: false,
                                 centered: @centered,
                                 background_color: @background_color,
-                                light_background_color: @light_background_color,
-                                dark_background_color: @dark_background_color,
+                                light_mode_background_color: @light_mode_background_color,
+                                dark_mode_background_color: @dark_mode_background_color,
                               },
                               action: {
                                 "message@window" => "onWindowMessage",